requirements.txt
gql[all]
beautifulsoup4==4.9.3
pytz==2021.1
python-dateutil==2.8.1
